Allows padding to be added by a keyboard
Instead of losing the screen when the keyboard opens, sometimes you
want to move the content of the page.

This allows you to use a plugin to do so by allowing you to specify the
current keyboard height.

Also, initial padding can now be different (like if you modify the
message bar).